requirements.txt
python-dotenv
pyaudio
openwakeword
#tflite-runtime
onnxruntime
numpy
pyyaml
wave